Refactor parameter list editor
