+++ studio/Cartfile.resolved
@@ -1,4 +1,4 @@
-git "file:///Users/devinabbott/Projects/Logic" "c1a150a9297010d741418a2e2298cee88b0c234d"
+git "file:///Users/devinabbott/Projects/Logic" "a6875af54bb9f2e374e613afc58c93bf6d389cf7"
 github "dabbott/ColorPicker" "c1e637d2af9813ddba6a9b8e5e51f164780981ba"
 github "dabbott/ControlledComponents" "264729ba7769e8fd2a66eb4fad63fd6348eae767"
 github "dabbott/FileTree" "9c31b1ddca54c9f1b8894bbd70de6d21132a8e67"
